+++ src/Eventus.Samples.Web/Eventus.Samples.Web.csproj
@@ -1,12 +1,12 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp1.1</TargetFramework>
+    <TargetFramework>net461</TargetFramework>
   </PropertyGroup>
 
-  <PropertyGroup>
+  <!--<PropertyGroup>
     <PackageTargetFallback>$(PackageTargetFallback);portable-net45+win8+wp8+wpa81;</PackageTargetFallback>
-  </PropertyGroup>
+  </PropertyGroup>-->
 
   <PropertyGroup>
     <UserSecretsId>aspnet-Eventus.Samples.Web-2ea9d849-4374-4ab5-84ec-65542e819077</UserSecretsId>
@@ -16,7 +16,7 @@
     <DocumentationFile>bin\Debug\eventus.xml</DocumentationFile>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="FluentEmail.Mailgun" Version="2.1.0" />
+    <PackageReference Include="mailgun_csharp" Version="0.6.0" />
     <PackageReference Include="MediatR" Version="3.0.1" />
     <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="2.0.0" />
     <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.0.0" />
