+++ src/Eventus.Samples.Web/Services/EmailOptions.cs
@@ -1,0 +1,11 @@
+﻿namespace Eventus.Samples.Web.Services
+{
+    public class EmailOptions
+    {
+        public string MailGunAPIKey { get; set; }
+
+        public string EmailDomain { get; set; }
+
+        public string EmailFrom { get; set; }
+    }
+}
