+++ src/Eventus.Samples.Web/Services/SmsOptions.cs
@@ -1,0 +1,11 @@
+﻿namespace Eventus.Samples.Web.Services
+{
+    public class SmsOptions
+    {
+        public string TwilioAccountSID { get; set; }
+
+        public string TwilioPhoneNumberFrom { get; set; }
+
+        public string TwilioAuthToken { get; set; }
+    }
+}
